Started reworking the camera
* focalLength now means the length from the sensor to the aperture
* changed aperture to be a float

+++ Code/SPath/src/kaloffl/spath/examples/Bunny.scala
@@ -46,7 +46,7 @@ object Bunny {
           position = Vec3d(0, 4.5, 9),
           forward = bunnyForward.normalize,
           up = bunnyTop,
-          aperture = 0.2,
-          focalLength = bunnyForward.length))))
+          aperture = 0.2f,
+          focalLength = bunnyForward.length.toFloat))))
   }
 }
