+++ Code/SPath/src/kaloffl/spath/examples/Scatter3.scala
@@ -64,7 +64,7 @@ object Scatter3 {
           position = Vec3d(0, 5, 13),
           forward = front,
           up = up,
-          aperture = 0.1,
-          focalLength = Vec3d(0, 2.5, 13).length))))
+          aperture = 0.1f,
+          focalLength = Vec3d(0, 2.5, 13).length.toFloat))))
   }
 }
